handle parsing of option values spanning into two lines
