+++ lib/bios/menus.py
@@ -25,6 +25,10 @@ def check_if_menu_line_is_an_option(line):
 
 @keyword("Extract strings from frame")
 def extract_strings_from_frame(text):
+    """
+    Extract a list of selectable strings from frame appearing
+    when selecting option from a list such as: Option <state>.
+    """
     inside_frame = False
     extracted_strings = []
     re_frame_start = r"^.*/-{3,}\\.*$"
@@ -63,3 +67,41 @@ def get_list_options(menu):
             result.append(option[1:-1].strip())
         return result
     return []
+
+
+@keyword("Get Value From Brackets")
+def get_value_from_brackets(text):
+    """
+    This keyword returns the current value stored in brackets
+    [] or <>.
+
+    We actually check for the first bracket only (and strip the
+    other one if it exists), as some values span over more than
+    one line. In such a case, only part of the option value (from
+    the first line) would be returned.
+    """
+    pattern = r"[\[<].*"
+
+    matches = re.findall(pattern, text)
+
+    if matches:
+        value = matches[0].strip("<>[]")
+    else:
+        value = None
+
+    return value
+
+
+@keyword("Check User Password Management menu default state")
+def check_user_password_management_default_state(entries):
+    status_present = False
+    change_option_present = False
+    for entry in entries:
+        if "Admin Password Status" in entry:
+            status_present = True
+            if not "Not Installed" in entry:
+                raise AssertionError("Admin Password Status should be not installed")
+        if "Change Admin Password" in entry:
+            change_option_present = True
+
+    return status_present and change_option_present
